Separate out SchemaPointers from StateUpdateSummary
