+++ radix-engine/src/transaction/mod.rs
@@ -1,9 +1,13 @@
 mod preview_executor;
+mod schema_pointers;
 mod state_update_summary;
+mod system_reader;
 mod transaction_executor;
 mod transaction_receipt;
 
 pub use preview_executor::*;
+pub use schema_pointers::*;
 pub use state_update_summary::*;
+pub use system_reader::*;
 pub use transaction_executor::*;
 pub use transaction_receipt::*;
