+++ radix-engine/src/transaction/state_update_summary.rs
@@ -1,25 +1,21 @@
-use radix_engine_common::data::scrypto::ScryptoDecode;
-use radix_engine_common::prelude::{scrypto_decode, scrypto_encode};
-use radix_engine_interface::api::ObjectModuleId;
-use radix_engine_interface::blueprints::package::{
-    BlueprintDefinition, BlueprintVersionKey, TypePointer, PACKAGE_BLUEPRINTS_PARTITION_OFFSET,
-};
+use radix_engine_common::prelude::{scrypto_decode};
 use radix_engine_interface::blueprints::resource::LiquidFungibleResource;
 use radix_engine_interface::data::scrypto::model::*;
 use radix_engine_interface::math::*;
 use radix_engine_interface::types::*;
 use radix_engine_interface::*;
 use radix_engine_store_interface::{
-    db_key_mapper::{DatabaseKeyMapper, MappedSubstateDatabase, SpreadPrefixKeyMapper},
+    db_key_mapper::{SpreadPrefixKeyMapper},
     interface::SubstateDatabase,
 };
 use sbor::rust::ops::AddAssign;
 use sbor::rust::prelude::*;
 
 use crate::system::node_modules::type_info::TypeInfoSubstate;
-use crate::system::system::{FieldSubstate, KeyValueEntrySubstate};
+use crate::system::system::FieldSubstate;
 use crate::track::TrackedSubstateValue;
 use crate::track::{TrackedNode, Write};
+use crate::transaction::SystemReader;
 
 #[derive(Default, Debug, Clone, ScryptoSbor)]
 pub struct StateUpdateSummary {
@@ -27,8 +23,6 @@ pub struct StateUpdateSummary {
     pub new_components: Vec<ComponentAddress>,
     pub new_resources: Vec<ResourceAddress>,
     pub new_vaults: Vec<InternalAddress>,
-    pub substate_schema_pointers:
-        IndexMap<NodeId, IndexMap<PartitionNumber, IndexMap<SubstateKey, SubstateSchemaPointer>>>,
     pub balance_changes: IndexMap<GlobalAddress, IndexMap<ResourceAddress, BalanceChange>>,
     /// This field accounts for Direct vault recalls (and the owner is not loaded during the transaction);
     pub direct_vault_updates: IndexMap<NodeId, IndexMap<ResourceAddress, BalanceChange>>,
@@ -61,8 +55,6 @@ impl StateUpdateSummary {
             }
         }
 
-        let substate_schema_pointers = SubstateSchemaMapper::new(substate_db, &updates).run();
-
         let (balance_changes, direct_vault_updates) =
             BalanceAccounter::new(substate_db, &updates).run();
 
@@ -71,7 +63,6 @@ impl StateUpdateSummary {
             new_components: new_components.into_iter().collect(),
             new_resources: new_resources.into_iter().collect(),
             new_vaults: new_vaults.into_iter().collect(),
-            substate_schema_pointers,
             balance_changes,
             direct_vault_updates,
         }
@@ -105,90 +96,6 @@ impl BalanceChange {
             BalanceChange::Fungible(..) => panic!("Not non fungible"),
             BalanceChange::NonFungible { removed, .. } => removed,
         }
-    }
-}
-
-#[derive(Debug, Clone, ScryptoSbor, PartialEq, Eq)]
-pub enum SubstateSchemaPointer {
-    KeyValueStore,
-    Object(PackageAddress, TypePointer),
-    TypeInfo,
-}
-
-/// Note that the implementation below assumes that substate owned objects can not be
-/// detached. If this changes, we will have to account for objects that are removed
-/// from a substate.
-pub struct SubstateSchemaMapper<'a, S: SubstateDatabase> {
-    system_reader: SystemReader<'a, S>,
-    tracked: &'a IndexMap<NodeId, TrackedNode>,
-}
-
-impl<'a, S: SubstateDatabase> SubstateSchemaMapper<'a, S> {
-    pub fn new(substate_db: &'a S, tracked: &'a IndexMap<NodeId, TrackedNode>) -> Self {
-        Self {
-            system_reader: SystemReader::new(substate_db, tracked),
-            tracked,
-        }
-    }
-
-    pub fn run(
-        &self,
-    ) -> IndexMap<NodeId, IndexMap<PartitionNumber, IndexMap<SubstateKey, SubstateSchemaPointer>>>
-    {
-        let mut substate_schema_pointers = index_map_new();
-        for (node_id, tracked_node) in self.tracked {
-            for (partition_num, tracked_partition) in &tracked_node.tracked_partitions {
-                for (_, tracked_substate) in &tracked_partition.substates {
-                    let partition_description =
-                        self.system_reader.partition_description(partition_num);
-                    let schema_pointer = match partition_description {
-                        SystemPartitionDescription::Module(module_id, offset) => (|| {
-                            let blueprint_id = if let ObjectModuleId::Main = module_id {
-                                let main_type_info =
-                                    self.system_reader.get_type_info(node_id).unwrap();
-                                match main_type_info {
-                                    TypeInfoSubstate::Object(info) => {
-                                        info.blueprint_info.blueprint_id
-                                    }
-                                    TypeInfoSubstate::KeyValueStore(..) => {
-                                        return SubstateSchemaPointer::KeyValueStore
-                                    }
-                                    _ => panic!("Unexpected Type Info {:?}", main_type_info),
-                                }
-                            } else {
-                                module_id.static_blueprint().unwrap()
-                            };
-
-                            let bp_def = self
-                                .system_reader
-                                .get_blueprint_definition(&blueprint_id)
-                                .unwrap();
-                            let type_pointer = bp_def
-                                .interface
-                                .state
-                                .get_type_pointer(&offset, &tracked_substate.substate_key)
-                                .unwrap();
-
-                            SubstateSchemaPointer::Object(
-                                blueprint_id.package_address,
-                                type_pointer,
-                            )
-                        })(
-                        ),
-                        SystemPartitionDescription::TypeInfo => SubstateSchemaPointer::TypeInfo,
-                    };
-
-                    substate_schema_pointers
-                        .entry(node_id.clone())
-                        .or_insert(index_map_new())
-                        .entry(partition_num.clone())
-                        .or_insert(index_map_new())
-                        .insert(tracked_substate.substate_key.clone(), schema_pointer);
-                }
-            }
-        }
-
-        substate_schema_pointers
     }
 }
 
@@ -447,118 +354,5 @@ impl<'a, S: SubstateDatabase> BalanceAccounter<'a, S> {
             }
         }
         .map(|x| (resource_address, x))
-    }
-}
-
-pub enum SystemPartitionDescription {
-    Module(ObjectModuleId, PartitionOffset),
-    TypeInfo,
-}
-
-pub struct SystemReader<'a, S: SubstateDatabase> {
-    substate_db: &'a S,
-    tracked: &'a IndexMap<NodeId, TrackedNode>,
-}
-
-impl<'a, S: SubstateDatabase> SystemReader<'a, S> {
-    pub fn new(substate_db: &'a S, tracked: &'a IndexMap<NodeId, TrackedNode>) -> Self {
-        Self {
-            substate_db,
-            tracked,
-        }
-    }
-
-    pub fn partition_description(
-        &self,
-        partition_num: &PartitionNumber,
-    ) -> SystemPartitionDescription {
-        if partition_num.ge(&MAIN_BASE_PARTITION) {
-            let partition_offset = PartitionOffset(partition_num.0 - MAIN_BASE_PARTITION.0);
-            SystemPartitionDescription::Module(ObjectModuleId::Main, partition_offset)
-        } else if partition_num.ge(&ROLE_ASSIGNMENT_BASE_PARTITION) {
-            let partition_offset =
-                PartitionOffset(partition_num.0 - ROLE_ASSIGNMENT_BASE_PARTITION.0);
-            SystemPartitionDescription::Module(ObjectModuleId::RoleAssignment, partition_offset)
-        } else if partition_num.ge(&ROYALTY_BASE_PARTITION) {
-            let partition_offset = PartitionOffset(partition_num.0 - ROYALTY_BASE_PARTITION.0);
-            SystemPartitionDescription::Module(ObjectModuleId::Royalty, partition_offset)
-        } else if partition_num.ge(&METADATA_BASE_PARTITION) {
-            let partition_offset = PartitionOffset(partition_num.0 - METADATA_BASE_PARTITION.0);
-            SystemPartitionDescription::Module(ObjectModuleId::Metadata, partition_offset)
-        } else if partition_num.eq(&TYPE_INFO_FIELD_PARTITION) {
-            SystemPartitionDescription::TypeInfo
-        } else {
-            panic!("Should not get here")
-        }
-    }
-
-    pub fn get_type_info(&self, node_id: &NodeId) -> Option<TypeInfoSubstate> {
-        let type_info: TypeInfoSubstate = self
-            .fetch_substate::<SpreadPrefixKeyMapper, TypeInfoSubstate>(
-                node_id,
-                TYPE_INFO_FIELD_PARTITION,
-                &TypeInfoField::TypeInfo.into(),
-            )?;
-
-        Some(type_info)
-    }
-
-    pub fn get_blueprint_definition(
-        &self,
-        blueprint_id: &BlueprintId,
-    ) -> Option<BlueprintDefinition> {
-        let bp_version_key = BlueprintVersionKey::new_default(blueprint_id.blueprint_name.clone());
-        let definition = self
-            .fetch_substate::<SpreadPrefixKeyMapper, KeyValueEntrySubstate<BlueprintDefinition>>(
-                blueprint_id.package_address.as_node_id(),
-                MAIN_BASE_PARTITION
-                    .at_offset(PACKAGE_BLUEPRINTS_PARTITION_OFFSET)
-                    .unwrap(),
-                &SubstateKey::Map(scrypto_encode(&bp_version_key).unwrap()),
-            )?;
-
-        definition.value
-    }
-
-    fn fetch_substate<M: DatabaseKeyMapper, D: ScryptoDecode>(
-        &self,
-        node_id: &NodeId,
-        partition_num: PartitionNumber,
-        key: &SubstateKey,
-    ) -> Option<D> {
-        // FIXME: explore if we can avoid loading from substate database
-        // - Part of the engine still reads/writes substates without touching the TypeInfo;
-        // - Track does not store the initial value of substate.
-
-        self.fetch_substate_from_state_updates::<M, D>(node_id, partition_num, key)
-            .or_else(|| self.fetch_substate_from_database::<M, D>(node_id, partition_num, key))
-    }
-
-    fn fetch_substate_from_database<M: DatabaseKeyMapper, D: ScryptoDecode>(
-        &self,
-        node_id: &NodeId,
-        partition_num: PartitionNumber,
-        key: &SubstateKey,
-    ) -> Option<D> {
-        self.substate_db
-            .get_mapped::<M, D>(node_id, partition_num, key)
-    }
-
-    fn fetch_substate_from_state_updates<M: DatabaseKeyMapper, D: ScryptoDecode>(
-        &self,
-        node_id: &NodeId,
-        partition_num: PartitionNumber,
-        key: &SubstateKey,
-    ) -> Option<D> {
-        self.tracked
-            .get(node_id)
-            .and_then(|tracked_node| tracked_node.tracked_partitions.get(&partition_num))
-            .and_then(|tracked_module| tracked_module.substates.get(&M::to_db_sort_key(key)))
-            .and_then(|tracked_key| {
-                tracked_key
-                    .substate_value
-                    .get()
-                    .map(|e| e.as_typed().unwrap())
-            })
     }
 }
