+++ radix-engine/src/transaction/transaction_receipt.rs
@@ -15,6 +15,7 @@ use radix_engine_interface::data::scrypto::ScryptoDecode;
 use radix_engine_interface::types::*;
 use sbor::representations::*;
 use utils::ContextualDisplay;
+use crate::transaction::SchemaPointers;
 
 #[derive(Debug, Clone, Default, ScryptoSbor)]
 pub struct ResourcesUsage {
@@ -71,6 +72,7 @@ pub struct CommitResult {
     pub fee_summary: FeeSummary,
     pub application_events: Vec<(EventTypeIdentifier, Vec<u8>)>,
     pub application_logs: Vec<(Level, String)>,
+    pub schema_pointers: SchemaPointers,
     /// Optional, only when `EnabledModule::ExecutionTrace` is ON.
     /// Mainly for transaction preview.
     pub execution_trace: TransactionExecutionTrace,
@@ -85,6 +87,7 @@ impl CommitResult {
             fee_summary: Default::default(),
             application_events: Default::default(),
             application_logs: Default::default(),
+            schema_pointers: Default::default(),
             execution_trace: Default::default(),
         }
     }
